fixes especially on votes and quorum
