+++ src/app/api/webhooks/multibaas/route.ts
@@ -2,10 +2,12 @@ import { governanceABI } from '@celo/abis';
 import { sql } from 'drizzle-orm';
 import { NextRequest } from 'next/server';
 import { createHmac } from 'node:crypto';
+import { writeFileSync } from 'node:fs';
 import { Event } from 'src/app/governance/events';
 import database from 'src/config/database';
-import { proposalsTable, votesTable } from 'src/db/schema';
+import { votesTable } from 'src/db/schema';
 import fetchHistoricalEventsAndSaveToDBProgressively from 'src/features/governance/fetchHistoricalEventsAndSaveToDBProgressively';
+import updateProposalsInDB from 'src/features/governance/updateProposalsInDB';
 import { handleProposalEvent } from 'src/features/governance/utils/events/proposal';
 import { handleVoteEvent } from 'src/features/governance/utils/events/vote';
 import { celoPublicClient } from 'src/utils/client';
@@ -39,61 +41,23 @@ export async function POST(request: NextRequest): Promise<Response> {
     return new Response(null, { status: 403 });
   }
 
+  const proposalIdsToUpdate: bigint[] = [];
   for (const {
     data: { event },
   } of body) {
     // NOTE: in theory we _could_ just insert the `event.rawFields` directly in the db...
-    await fetchHistoricalEventsAndSaveToDBProgressively(event.name, celoPublicClient);
+    await fetchHistoricalEventsAndSaveToDBProgressively(event.name!, celoPublicClient);
 
     const eventData: Event = JSON.parse(event.rawFields);
 
     // NOTE: for clarity, we don't need to parallelize `handleXXXEvent`
-    // sine they just exit early when the event doesnt match
+    // since they just exit early when the event doesnt match
+    const proposalId = await handleProposalEvent(event.name!, eventData);
+    if (proposalId) {
+      proposalIdsToUpdate.push(proposalId);
+    }
 
-    await handleProposalEvent(event.name!, eventData, event.inputs, celoPublicClient).then(
-      async (proposal) => {
-        if (!proposal) return;
-
-        await database
-          .insert(proposalsTable)
-          .values(proposal)
-          .onConflictDoUpdate({
-            set: {
-              // NOTE: normally only the stage and neworkWeight should be updated over time...
-              stage: sql`excluded."stage"`,
-              networkWeight: sql`excluded."networkWeight"`,
-              // NOTE: but maybe other things were updated in github?
-              author: sql`excluded."author"`,
-              url: sql`excluded."url"`,
-              cgpUrl: sql`excluded."cgpUrl"`,
-              cgpUrlRaw: sql`excluded."cgpUrlRaw"`,
-              title: sql`excluded."title"`,
-            },
-            target: [proposalsTable.chainId, proposalsTable.id],
-          });
-
-        console.info(`Upserted proposal ${proposal.id} to stage ${proposal.stage}`);
-
-        const allProposals = await database
-          .select({
-            id: proposalsTable.id,
-          })
-          .from(proposalsTable)
-          .where(sql`${proposalsTable.cgp} = ${proposal.cgp}`)
-          .groupBy(proposalsTable.cgp);
-
-        const ids = allProposals.map((x) => x.id);
-        while (ids.length > 1) {
-          const [pastId] = ids.splice(0, 1);
-          await database
-            .update(proposalsTable)
-            .set({ pastId: pastId })
-            .where(sql`${proposalsTable.id} = ${ids.at(0)}`);
-        }
-      },
-    );
-
-    await handleVoteEvent(event.name!, eventData, celoPublicClient).then(async (rows) => {
+    await handleVoteEvent(event.name!, eventData, celoPublicClient.chain.id).then(async (rows) => {
       if (!rows.length) return;
 
       const { count } = await database
@@ -105,6 +69,10 @@ export async function POST(request: NextRequest): Promise<Response> {
         });
       console.info(`Inserted ${count} vote records for proposal: ${rows[0].proposalId}`);
     });
+  }
+
+  if (proposalIdsToUpdate.length) {
+    await updateProposalsInDB(celoPublicClient, proposalIdsToUpdate);
   }
 
   return new Response(null, { status: 200 });
@@ -131,6 +99,16 @@ function assertSignature(
       signature,
       signature_,
     });
+    writeFileSync(
+      '/Users/nicolasbrugneaux/workspace/celo-mondo/logs/' + Date.now() + '-error.json',
+      JSON.stringify({
+        payload,
+        signature,
+        signature_,
+        timestamp,
+        secret: process.env.MULTIBAAS_WEBHOOK_SECRET!,
+      }),
+    );
     return false;
   }
   return JSON.parse(payload);
