+++ src/features/governance/hooks/useGovernanceProposals.ts
@@ -1,11 +1,13 @@
+import { governanceABI } from '@celo/abis';
 import { useQuery } from '@tanstack/react-query';
 import { useMemo } from 'react';
 import { useToastError } from 'src/components/notifications/useToastError';
 import { GCTime, StaleTime } from 'src/config/consts';
+import { Addresses } from 'src/config/contracts';
 import { fetchProposalsFromRepo } from 'src/features/governance/fetchFromRepository';
 import { fetchProposals } from 'src/features/governance/fetchProposals';
 import { getExpiryTimestamp, MergedProposalData } from 'src/features/governance/governanceData';
-import { ProposalMetadata, ProposalStage } from 'src/features/governance/types';
+import { ProposalMetadata, ProposalStage, VoteAmounts } from 'src/features/governance/types';
 import { logger } from 'src/utils/logger';
 import { usePublicClient } from 'wagmi';
 
@@ -16,10 +18,8 @@ export function useGovernanceProposal(id?: number) {
 }
 
 function useGovernanceDrafts() {
-  const publicClient = usePublicClient();
-
   const { isLoading, isError, error, data } = useQuery({
-    queryKey: ['useGovernanceDrafts', publicClient],
+    queryKey: ['useGovernanceDrafts'],
     queryFn: async () => {
       logger.debug('Fetching governance drafts');
 
@@ -55,38 +55,59 @@ export function useGovernanceProposals() {
       logger.debug('Fetching governance proposals');
       // Fetch on-chain data
       const proposals = await fetchProposals(publicClient.chain!.id);
-      return proposals.map((proposal) => ({
-        id: proposal.id,
-        stage: proposal.stage,
-        history: proposal.history,
-        metadata: {
-          author: proposal.author,
-          cgp: proposal.cgp,
-          cgpUrl: proposal.cgpUrl,
-          cgpUrlRaw: proposal.cgpUrlRaw,
-          stage: proposal.stage,
-          title: proposal.title,
-          timestamp: proposal.timestamp * 1000,
-          timestampExecuted: proposal.executedAt ? proposal.executedAt * 1000 : null,
+      const [ids, upvotesArr] = await publicClient.readContract({
+        address: Addresses.Governance,
+        abi: governanceABI,
+        functionName: 'getQueue',
+      });
+
+      return proposals.map((proposal) => {
+        const queuedId = ids.indexOf(BigInt(proposal.id));
+        let upvotes = 0n;
+        if (queuedId !== -1) {
+          upvotes = upvotesArr.at(queuedId)!;
+        }
+
+        // NOTE: for some reason nested bigint are deserialized?
+        Object.keys(proposal.votes).forEach((voteType) => {
+          proposal.votes[voteType as keyof VoteAmounts] = BigInt(
+            proposal.votes[voteType as keyof VoteAmounts],
+          );
+        });
+
+        return {
           id: proposal.id,
-          url: proposal.url,
-          votes: proposal.votes,
-        },
-        proposal: {
-          deposit: proposal.deposit,
-          id: proposal.id,
-          networkWeight: proposal.networkWeight,
-          numTransactions: BigInt(proposal.transactionCount || 0),
           stage: proposal.stage,
-          proposer: proposal.proposer,
-          upvotes: 0n, // TODO
-          url: proposal.url,
-          expiryTimestamp: getExpiryTimestamp(proposal.stage, proposal.timestamp * 1000),
-          timestamp: proposal.timestamp * 1000,
-          isApproved: proposal.stage > ProposalStage.Approval,
-          votes: proposal.votes,
-        },
-      })) as MergedProposalData[];
+          history: proposal.history,
+          metadata: {
+            author: proposal.author,
+            cgp: proposal.cgp,
+            cgpUrl: proposal.cgpUrl,
+            cgpUrlRaw: proposal.cgpUrlRaw,
+            stage: proposal.stage,
+            title: proposal.title,
+            timestamp: proposal.timestamp * 1000,
+            timestampExecuted: proposal.executedAt ? proposal.executedAt * 1000 : null,
+            id: proposal.id,
+            url: proposal.url,
+            votes: proposal.votes,
+          },
+          proposal: {
+            deposit: proposal.deposit,
+            id: proposal.id,
+            networkWeight: proposal.networkWeight,
+            numTransactions: BigInt(proposal.transactionCount || 0),
+            stage: proposal.stage,
+            proposer: proposal.proposer,
+            upvotes,
+            url: proposal.url,
+            expiryTimestamp: getExpiryTimestamp(proposal.stage, proposal.timestamp * 1000),
+            timestamp: proposal.timestamp * 1000,
+            isApproved: proposal.stage > ProposalStage.Approval,
+            votes: proposal.votes,
+          },
+        } as MergedProposalData;
+      });
     },
     gcTime: GCTime.Long,
     staleTime: StaleTime.Default,
